Document extractor CLI in README
