+++ extractor/__init__.py
@@ -1,0 +1,5 @@
+"""Utilities for extracting policy document text artifacts."""
+
+__all__ = ["__version__"]
+
+__version__ = "0.1.0"
